Move more into lib/
