+++ lib/hierarchy.js
