+++ lib/infer/kind.js
