+++ lib/infer/membership.js
