+++ lib/nest_params.js
@@ -12,7 +12,8 @@ var extend = require('extend');
  * This stream assumes that incoming comments have been flattened.
  *
  * @name nestParams
- * @return {stream.Transform}
+ * @param {Object} comment input comment
+ * @return {Object} nested comment
  */
 module.exports = function (comment) {
   if (!comment.params) {
