+++ lib/parsers/javascript.js
