+++ lib/parsers/polyglot.js
@@ -1,0 +1,27 @@
+'use strict';
+
+var getComments = require('get-comments'),
+  extend = require('extend'),
+  isJSDocComment = require('../../lib/is_jsdoc_comment'),
+  parse = require('../../lib/parse');
+
+/**
+ * Documentation stream parser: this receives a module-dep item,
+ * reads the file, parses the JavaScript, parses the JSDoc, and
+ * emits parsed comments.
+ * @name parse
+ * @param {Object} data a chunk of data provided by module-deps
+ * @return {Array<Object>} adds to memo
+ */
+module.exports = function (memo, data) {
+  getComments(data.source, true)
+    .filter(isJSDocComment)
+    .forEach(function (comment) {
+      var context = {
+        loc: extend({}, comment.loc),
+        file: data.file
+      };
+      memo.push(parse(comment.value, comment.loc, context));
+    });
+  return memo;
+};
