+++ streams/output/html.js
@@ -5,7 +5,7 @@ var File = require('vinyl'),
   Handlebars = require('handlebars'),
   extend = require('extend'),
   slugg = require('slugg'),
-  hierarchy = require('../hierarchy'),
+  hierarchy = require('../../lib/hierarchy'),
   getTemplate = require('./lib/get_template'),
   resolveTheme = require('./lib/resolve_theme'),
   helpers = require('./lib/html_helpers'),
